Move generation and tests to clients/python
